requirements.txt
aiodns==3.0.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
alluka==0.1.3
async-timeout==4.0.2
attrs==23.2.0
blinker==1.7.0
Brotli==1.0.9
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.0.1
ciso8601==2.3.0
click==8.1.7
colorama==0.4.6
colorlog==6.7.0
Flask==3.0.2
frozenlist==1.3.3
hikari==2.0.0.dev121
hikari-miru==3.1.2
hikari-tanjun==2.17.2
humanfriendly==10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.4
orjson==3.8.6
pet-pet-gif==1.0.2
Pillow==9.4.0
pip==24.0
pycares==4.3.0
pycparser==2.21
PyNaCl==1.5.0
pyreadline3==3.4.1
python-dotenv==0.21.1
requests==2.31.0
setuptools==69.1.0
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.8.2
